+++ apps/widgetbook_samples/lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+import 'package:get_it/get_it.dart';
 import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
 import 'package:widgetbook_samples/l10n/app_localizations.dart';
@@ -6,6 +8,9 @@ import 'package:widgetbook_samples/l10n/app_localizations.dart';
 import 'main.directories.g.dart';
 
 void main() async {
+  GetIt.instance.registerLazySingleton<BaseCacheManager>(
+    () => DefaultCacheManager(),
+  );
   runApp(const WidgetbookApp());
 }