+++ packages/widgetbook_golden_test/test/integration/run_widgetbook_golden_tests_test.dart
@@ -1,10 +1,12 @@
-import 'package:widgetbook_golden_test/src/run_widgetbook_golden_tests.dart';
-import 'package:widgetbook_golden_test_core/src/widgetbook_golden_tests_properties.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+import 'package:get_it/get_it.dart';
+import 'package:widgetbook_golden_test/widgetbook_golden_test.dart';
 import 'package:widgetbook_samples/main.directories.g.dart';
 
 void main() {
-  runWidgetbookGoldenTests(
-    nodes: directories,
-    properties: WidgetbookGoldenTestsProperties(),
+  final properties = WidgetbookGoldenTestsProperties();
+  GetIt.instance.registerLazySingleton<BaseCacheManager>(
+    () => MockTestCacheManager(properties: properties),
   );
+  runWidgetbookGoldenTests(nodes: directories, properties: properties);
 }